+++ app/controllers/providers/means/check_answers_incomes_controller.rb
@@ -1,0 +1,27 @@
+module Providers
+  module Means
+    class CheckAnswersIncomesController < ProviderBaseController
+      before_action :set_transaction_types, only: :show
+
+      def show
+        legal_aid_application.set_transaction_period
+      end
+
+      def update
+        continue_or_draft
+      end
+
+    private
+
+      def set_transaction_types
+        @credit_transaction_types = if legal_aid_application.uploading_bank_statements?
+                                      TransactionType.credits.without_disregarded_benefits
+                                    else
+                                      TransactionType.credits.without_housing_benefits
+                                    end
+
+        @debit_transaction_types = TransactionType.debits
+      end
+    end
+  end
+end
