+++ app/controllers/providers/means_summaries_controller.rb
@@ -1,9 +1,9 @@
 module Providers
   class MeansSummariesController < ProviderBaseController
-    before_action :set_transaction_types, only: :show
+    # before_action :set_transaction_types, only: :show
 
     def show
-      legal_aid_application.set_transaction_period
+      # legal_aid_application.set_transaction_period
       legal_aid_application.check_non_passported_means! unless legal_aid_application.checking_non_passported_means?
     end
 
@@ -32,14 +32,14 @@ module Providers
       CFE::SubmissionManager.call(legal_aid_application.id)
     end
 
-    def set_transaction_types
-      @credit_transaction_types = if legal_aid_application.uploading_bank_statements?
-                                    TransactionType.credits.without_disregarded_benefits
-                                  else
-                                    TransactionType.credits.without_housing_benefits
-                                  end
-
-      @debit_transaction_types = TransactionType.debits
-    end
+    # def set_transaction_types
+    #   @credit_transaction_types = if legal_aid_application.uploading_bank_statements?
+    #                                 TransactionType.credits.without_disregarded_benefits
+    #                               else
+    #                                 TransactionType.credits.without_housing_benefits
+    #                               end
+    #
+    #   @debit_transaction_types = TransactionType.debits
+    # end
   end
 end
