+++ app/views/providers/means/check_answers_incomes/_income_assessment.html.erb
@@ -1,6 +1,6 @@
 <% bank_statement_upload = @legal_aid_application.uploading_bank_statements? %>
 
-<h2 class="govuk-heading-l"><%= t(".income-heading") %></h2>
+<h2 class="govuk-heading-l govuk-!-margin-bottom-8"><%= t('.income-heading') %></h2>
 
 <% if bank_statement_upload %>
   <%= render("shared/check_answers/bank_statements", bank_statements: @legal_aid_application.attachments.bank_statement_evidence, read_only: false) %>
@@ -27,7 +27,11 @@
 
 <%= render "shared/check_answers/student_finance", read_only: false %>
 
+<<<<<<< HEAD:app/views/providers/means_summaries/_income_assessment.html.erb
 <h2 class="govuk-heading-l"><%= t(".outgoings-heading") %></h2>
+=======
+<h2 class="govuk-heading-l govuk-!-margin-top-8 govuk-!-margin-bottom-8"><%= t('.outgoings-heading') %></h2>
+>>>>>>> add new check answers page for income section:app/views/providers/means/check_answers_incomes/_income_assessment.html.erb
 
 <%= render(
       "shared/check_answers/income_summary",
@@ -40,6 +44,10 @@
 
 <%= render("shared/check_answers/outgoings_cash_payments", read_only: false) if @legal_aid_application.uploading_bank_statements? %>
 
+<<<<<<< HEAD:app/views/providers/means_summaries/_income_assessment.html.erb
 <h2 class="govuk-heading-l"><%= t(".dependants.heading") %></h2>
+=======
+<h2 class="govuk-heading-l"><%= t('.dependants-heading') %></h2>
+>>>>>>> add new check answers page for income section:app/views/providers/means/check_answers_incomes/_income_assessment.html.erb
 
 <%= render "shared/check_answers/dependants" %>
