+++ app/views/providers/means_summaries/show.html.erb
@@ -1,7 +1,5 @@
 <%= page_template(page_title: t(".h1-heading")) do %>
 
-  <%= render "providers/means_summaries/income_assessment" %>
-
   <h2 class="govuk-heading-l"><%= t(".capital-section-heading") %></h2>
 
   <%= render "shared/check_answers/assets" %>
