+++ config/locales/en/providers.yml
@@ -884,6 +884,18 @@ en:
         show:
           page_title: "Are you sure you want to remove %{name}?"
           error: Select yes if you wish to delete the dependant
+      check_answers_incomes:
+        show:
+          h1-heading: Check your answers
+          what_happens_next:
+            heading: What happens next
+            text: We'll ask you questions about your client's capital
+        income_assessment:
+          income-heading: Your client's income
+          outgoings-heading: Your client's outgoings
+          credits-section-heading: What payments does your client receive?
+          debits-section-heading: What payments does your client make?
+          dependants-heading: Dependants
       own_homes:
         show:
           h1-heading: Does your client own the home that they live in?
@@ -956,23 +968,6 @@ en:
         what_happens_next:
           heading: What happens next
           text: We'll use your answers to check your client's financial eligibility for legal aid.
-      income_assessment:
-        income-heading: Your client's income
-        outgoings-heading: Your client's outgoings
-        dependants:
-          heading: Dependants
-          has_dependants: Does your client have any dependants?
-          add-new: Add dependant
-          delete: Delete dependant
-          name: Name
-          date_of_birth: Date of birth
-          relationship: What is their relationship to your client?
-          in_full_time_education: Are they in full-time education or training?
-          income: Do they receive any income?
-          assets: Do they have assets worth more than £8,000?
-        credits-section-heading: What payments does your client receive?
-        debits-section-heading: What payments does your client make?
-
     merits_task_lists:
       show:
         page_heading: Provide details of the case
